+++ lib/testing/testCHORDTelescope.cpp
@@ -40,7 +40,9 @@ REGISTER_KOTEKAN_STAGE(TestCHORDTelescope);
 TestCHORDTelescope::TestCHORDTelescope(Config& config, const std::string& unique_name,
                                        bufferContainer& buffer_container) :
     Stage(config, unique_name, buffer_container,
-          std::bind(&TestCHORDTelescope::main_thread, this)) {}
+          std::bind(&TestCHORDTelescope::main_thread, this)),
+    do_dishes(config.get_default<bool>(unique_name, "do_dishes", true)),
+    do_eop_probes(config.get_default<bool>(unique_name, "do_eop_probes", true)) {}
 
 TestCHORDTelescope::~TestCHORDTelescope() {}
 
@@ -93,7 +95,7 @@ void TestCHORDTelescope::main_thread() {
         for (i = 0; i < eop_tab.size(); i++) {
             struct EOP eop = eop_tab[i];
             eop_times.push_back(eop.t_inst);
-            INFO("            {0: 2d} - t_inst: {1:d} s + {2:d} ns", i, eop.t_inst / GIGA,
+            INFO("            {0:02d} - t_inst: {1:d} s + {2:d} ns", i, eop.t_inst / GIGA,
                  eop.t_inst % GIGA);
             INFO("               - t_ut1:  {0:d} s + {1:d} ns", eop.t_ut1 / GIGA, eop.t_ut1 % GIGA);
             INFO("               - dut1:   {:f} s", eop.delta_UT1_inst);
@@ -104,79 +106,83 @@ void TestCHORDTelescope::main_thread() {
 
         size_t nt = eop_times.size();
 
-        INFO("            EOP Probes:");
+        if (do_eop_probes) {
+            INFO("            EOP Probes:");
 
-        for (i = 0; i <= nt; i++) {
-            int64_t ta, tb;
-            if (i == 0 && nt == 1)
-                ta = eop_times[0] - 43200 * GIGA;
-            else if (i == 0)
-                ta = eop_times[0] - 2 * (eop_times[1] - eop_times[0]);
-            else
-                ta = eop_times[i - 1];
+            for (i = 0; i <= nt; i++) {
+                int64_t ta, tb;
+                if (i == 0 && nt == 1)
+                    ta = eop_times[0] - 43200 * GIGA;
+                else if (i == 0)
+                    ta = eop_times[0] - 2 * (eop_times[1] - eop_times[0]);
+                else
+                    ta = eop_times[i - 1];
 
-            if (i == nt && nt == 1)
-                tb = eop_times[0] + 43200 * GIGA;
-            else if (i == nt)
-                tb = eop_times[nt - 1] + 2 * (eop_times[nt - 1] - eop_times[nt - 2]);
-            else
-                tb = eop_times[i];
-
-
-            int n_seg = 4;
-            int64_t dns = tb - ta;
-            for (int j = 0; j < n_seg; j++) {
-
-                int64_t t = ta + (j * dns) / n_seg;
-                timespec ts = {.tv_sec = (time_t)(t / GIGA), .tv_nsec = t % GIGA};
-
-                struct EOP eop = tel.get_EOP_at_time(ts);
-                INFO("               - t_inst: {1:d} s + {2:d} ns", i, eop.t_inst / GIGA,
-                     eop.t_inst % GIGA);
-                INFO("               - t_ut1:  {0:d} s + {1:d} ns", eop.t_ut1 / GIGA,
-                     eop.t_ut1 % GIGA);
-                INFO("               - dut1:   {:f} s", eop.delta_UT1_inst);
-                INFO("               - era:    {:f} deg", eop.ERA_deg);
-                INFO("               - xp:     {:f} arcsec", eop.xp_as);
-                INFO("               - yp:     {:f} arcsec", eop.yp_as);
-
-                timespec ts_inst2 = get_time_from_UT1(eop.t_ut1, eop.delta_UT1_inst);
-                int64_t n_rot;
-                double era = get_ERA_from_UT1(eop.t_ut1, &n_rot);
-                int64_t t_ut12 = get_UT1_from_ERA(n_rot, eop.ERA_deg);
-
-                INFO("               -t_inst2: {0:d} s + {1:d} ns", ts_inst2.tv_sec,
-                     ts_inst2.tv_nsec);
-                INFO("               -diff:    {0:d} s + {1:d} ns",
-                     ts_inst2.tv_sec - eop.t_inst / GIGA, ts_inst2.tv_nsec - eop.t_inst % GIGA);
-                INFO("               -t_ut12:  {0:d} s + {1:d} ns", t_ut12 / GIGA, t_ut12 % GIGA);
-                INFO("               -diff:    {0:d} s + {1:d} ns", (t_ut12 - eop.t_ut1) / GIGA,
-                     (t_ut12 - eop.t_ut1) % GIGA);
+                if (i == nt && nt == 1)
+                    tb = eop_times[0] + 43200 * GIGA;
+                else if (i == nt)
+                    tb = eop_times[nt - 1] + 2 * (eop_times[nt - 1] - eop_times[nt - 2]);
+                else
+                    tb = eop_times[i];
 
 
-                int64_t n_rot2;
-                double era2 = get_ERA_from_UT1(t_ut12, &n_rot2);
+                int n_seg = 4;
+                int64_t dns = tb - ta;
+                for (int j = 0; j < n_seg; j++) {
 
-                INFO("               -era:  {0:f} deg + {1:d}", era, n_rot);
-                INFO("               -diff: {0:e} deg + {1:d}", era2 - era, n_rot2 - n_rot);
+                    int64_t t = ta + (j * dns) / n_seg;
+                    timespec ts = {.tv_sec = (time_t)(t / GIGA), .tv_nsec = t % GIGA};
+
+                    struct EOP eop = tel.get_EOP_at_time(ts);
+                    INFO("               - t_inst: {1:d} s + {2:d} ns", i, eop.t_inst / GIGA,
+                         eop.t_inst % GIGA);
+                    INFO("               - t_ut1:  {0:d} s + {1:d} ns", eop.t_ut1 / GIGA,
+                         eop.t_ut1 % GIGA);
+                    INFO("               - dut1:   {:f} s", eop.delta_UT1_inst);
+                    INFO("               - era:    {:f} deg", eop.ERA_deg);
+                    INFO("               - xp:     {:f} arcsec", eop.xp_as);
+                    INFO("               - yp:     {:f} arcsec", eop.yp_as);
+
+                    timespec ts_inst2 = get_time_from_UT1(eop.t_ut1, eop.delta_UT1_inst);
+                    int64_t n_rot;
+                    double era = get_ERA_from_UT1(eop.t_ut1, &n_rot);
+                    int64_t t_ut12 = get_UT1_from_ERA(n_rot, eop.ERA_deg);
+
+                    INFO("               -t_inst2: {0:d} s + {1:d} ns", ts_inst2.tv_sec,
+                         ts_inst2.tv_nsec);
+                    INFO("               -diff:    {0:d} s + {1:d} ns",
+                         ts_inst2.tv_sec - eop.t_inst / GIGA, ts_inst2.tv_nsec - eop.t_inst % GIGA);
+                    INFO("               -t_ut12:  {0:d} s + {1:d} ns", t_ut12 / GIGA, t_ut12 % GIGA);
+                    INFO("               -diff:    {0:d} s + {1:d} ns", (t_ut12 - eop.t_ut1) / GIGA,
+                         (t_ut12 - eop.t_ut1) % GIGA);
+
+
+                    int64_t n_rot2;
+                    double era2 = get_ERA_from_UT1(t_ut12, &n_rot2);
+
+                    INFO("               -era:  {0:f} deg + {1:d}", era, n_rot);
+                    INFO("               -diff: {0:e} deg + {1:d}", era2 - era, n_rot2 - n_rot);
+                }
             }
         }
 
-        size_t n_dish = tel.get_num_dishes();
-        INFO("            Num Dishes:  {:d}", n_dish);
-        for (i = 0; i < n_dish; i++) {
-            std::array<double, 3> pos = tel.get_dish_position_in_grid_coords(i);
-            INFO("            Dish Pos:    {0:d} - ({1:f}, {2:f}, {3:f})", i, pos[0], pos[1],
-                 pos[2]);
-        }
+        if (do_dishes) {
+            size_t n_dish = tel.get_num_dishes();
+            INFO("            Num Dishes:  {:d}", n_dish);
+            for (i = 0; i < n_dish; i++) {
+                std::array<double, 3> pos = tel.get_dish_position_in_grid_coords(i);
+                INFO("            Dish Pos:    {0:d} - ({1:f}, {2:f}, {3:f})", i, pos[0], pos[1],
+                     pos[2]);
+            }
 
-        for (i = 0; i < n_dish; i++) {
-            json j = tel.get_dish_at_idx(i);
-            INFO("            Dish Info: {:s}", j.dump());
+            for (i = 0; i < n_dish; i++) {
+                json j = tel.get_dish_at_idx(i);
+                INFO("            Dish Info: {:s}", j.dump());
+            }
         }
 
         // break;
-        std::this_thread::sleep_for(std::chrono::milliseconds(2000));
+        std::this_thread::sleep_for(std::chrono::milliseconds(10'000));
     }
 
     exit_kotekan(CLEAN_EXIT);
