+++ lib/testing/testCHORDTelescope.hpp
@@ -37,6 +37,9 @@ public:
      * @brief Framework managed pthread.
      */
     void main_thread() override;
+private:
+    const bool do_dishes;
+    const bool do_eop_probes;
 };
 
 #endif /* EXAMPLE_CONSUMER_H */
